refactor: extract HealthService from McpProxy
Extract health reporting into a dedicated HealthService class:
- BuildHealthReport(), BuildHealthToolResponse()
- HealthTool static Tool definition
- HealthResourceUri constant

McpProxy now delegates all health logic to HealthService via DI.
